# Conflicts:
#	app/controllers/ProductController.java
#	app/models/dao/ProductDAO.java
#	conf/routes

+++ app/controllers/ProductController.java
@@ -15,12 +15,19 @@ import play.mvc.Controller;
 import play.mvc.Result;
 import play.db.jpa.Transactional;
 import play.mvc.Security;
+
 import java.util.*;
+
 import javax.inject.Inject;
+
+import static utils.LinkParser.*;
+
 
 /**
  * Created by octavian.salcianu on 7/14/2016.
  */
+
+
 public class ProductController extends Controller {
 	@Inject
 	private ProductDAO productDAO;
@@ -42,38 +49,39 @@ public class ProductController extends Controller {
 		} else {
 			JsonNode json = request().body().asJson();
 			Form<Product> form = formFactory.form(Product.class).bind(json);
-			if (form.hasErrors()) {
+			if (form.hasErrors())
 				return badRequest("Invalid form");
-			}
-			String smth = form.get().getLinkAddress().split("/")[0];
 			Product product = new Product();
 
-			if((form.get().getProdName()) != null && (form.get().getLinkAddress() != null) && (smth != null)){
-				product = form.get();
-			} else {
-				return badRequest("Invalid form");
-			}
-			Site site = siteDAO.getSiteByURL(product.getLinkAddress().split("/")[0]);
-			if (site != null) {
+			product = form.get();
+
+			Site site = siteDAO.getSiteByURL(parseSite(product.getLinkAddress()));
+
+			if(site != null)
 				product.setSite(site);
-			} else {
-				return notFound("No such site");
-			}
-			/* I'm aware of this code duplication, will be fixed after release */
-			String URL = product.getLinkAddress();
-			String[] URLsite = URL.split("/");
-			String[] URLkeywords = URLsite[1].split("-");
+			else
+				return badRequest("No such site");
+
+			String[] keywords = parseKeywordsFromLink(product.getLinkAddress());
+			if(keywords == null)
+				return badRequest("Invalid site or missing meta tag");
+
+			if(keywords.length == 1 && keywords[0].equals("getFromName"))
+				keywords = parseKeywordsFromName(product.getProdName());
+
 			Set<Keyword> kk = new HashSet<>();
-			for(String s : URLkeywords){
+
+			/* Create keyword objects */
+			for(String kw : keywords){
 				Keyword tibi = new Keyword();
 				tibi.setId(null);
 				tibi.setProduct(product);
-				tibi.setKeyword(s);
+				tibi.setKeyword(kw);
 				kk.add(tibi);
 			}
 			product.setKeywords(kk);
 			productDAO.create(product);
-			return ok("Added");
+			return ok("Added product: " + product.getProdName());
 		}
 	}
 
@@ -88,11 +96,12 @@ public class ProductController extends Controller {
 				return notFound("No such product");
 			} else{
 				/* Hard delete */
-//				productDAO.delete(product);
+				productDAO.delete(product);
 
 				/* Soft delete */
-				productDAO.softDelete(product);
-				return ok("Deleted");
+//				productDAO.softDelete(product);
+				return ok("Product deleted: " + product.getProdName());
+
 			}
 		}
 	}
@@ -106,49 +115,54 @@ public class ProductController extends Controller {
 		} else {
 			JsonNode json = request().body().asJson();
 			Form<Product> form = formFactory.form(Product.class).bind(json);
-
 			if (form.hasErrors()) {
 				return badRequest("Invalid form");
 			}
-
 			Product current = new Product();
-
 			if(id != null){
 				current = productDAO.get(id);
 			} else{
 				return notFound("Pls provide ID!!!");
 			}
-
 			if (current == null) {
 				return notFound("Product doesn't exist");
 			} else {
+
 				if(form.get().getLinkAddress().equalsIgnoreCase(current.getLinkAddress())){
 					current.setProdName(form.get().getProdName());
 					current.setLinkAddress(form.get().getLinkAddress());
-					keywordDAO.delete(current);
 					productDAO.update(current);
 				} else {
 					current.setProdName(form.get().getProdName());
 					current.setLinkAddress(form.get().getLinkAddress());
-					/* I'm aware of this code duplication, will be fixed after release */
-					String URL = current.getLinkAddress();
-					String[] URLsite = URL.split("/");
-					String[] URLkeywords = URLsite[1].split("-");
-					Set<Keyword> kk = new HashSet<>();
 
-					for(String s : URLkeywords){
+					Site site = siteDAO.getSiteByURL(parseSite(current.getLinkAddress()));
+
+					if(site != null)
+						current.setSite(site);
+					else
+						return badRequest("No such site");
+
+					String[] keywords = parseKeywordsFromLink(current.getLinkAddress());
+					if(keywords == null)
+						return badRequest("Invalid site or missing meta tag");
+
+					if(keywords.length == 1 && keywords[0].equals("getFromName"))
+						keywords = parseKeywordsFromName(current.getProdName());
+
+					keywordDAO.delete(current);
+					Set<Keyword> kk = new HashSet<>();
+					for(String s : keywords){
 						Keyword tibi = new Keyword();
 						tibi.setId(null);
 						tibi.setProduct(current);
 						tibi.setKeyword(s);
 						kk.add(tibi);
 					}
-
 					current.setKeywords(kk);
 					productDAO.update(current);
 				}
-
-				return ok("Success");
+				return ok("Product updated: " + current.getProdName());
 			}
 		}
 	}
